requirements.txt
# Core Data Science & ML
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.0.0
scipy>=1.0.0
Pillow>=10.0.0

# Deep Learning & AI
torch>=2.0.0
transformers>=4.53.3
peft>=0.12.0
accelerate>=1.9.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Vector Database & Embeddings
faiss-cpu==1.8.0
sentence-transformers==3.0.1
FlagEmbedding>=1.3.0

# LangChain & AI Frameworks
langchain==0.2.16
langchain_community==0.2.16
langchain_huggingface==0.0.3
langchain_core==0.2.38
groq==0.9.0
hugchat==0.4.0
google-generativeai>=0.8.0
huggingface-hub>=0.15.0

# Document Processing
pypdf==4.3.1
python-docx==1.1.2

# Web Framework - Flask
flask>=3.1.0
flask-cors==4.0.1
flask-restful==0.3.10
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-bcrypt==1.0.1
flask-wtf==1.2.1
flask-mail==0.10.0
werkzeug>=3.1.3

# Database
mysqlclient==2.2.4
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.40

# Streamlit
streamlit==1.37.1

# Utilities
python-dotenv==1.0.1
email-validator>=2.2.0
requests>=2.32.3
tqdm>=4.65.0
pyyaml==6.0.2

# Production Server
gunicorn>=23.0.0 